Avoid some warnings.

+++ WinArk/DPCTimerTable.cpp
@@ -162,7 +162,7 @@ void CDpcTimerTable::Refresh() {
 			if (p != nullptr) {
 				memset(p, 0, size);
 				DriverHelper::EnumKernelTimer(&data, p, size);
-				for (int i = 0; i < count; i++) {
+				for (ULONG i = 0; i < count; i++) {
 					std::shared_ptr<DpcTimerInfo> info = std::make_shared<DpcTimerInfo>();
 					info->DueTime = p[i].DueTime;
 					info->KDpc = p[i].KDpc;
